Large refactor of physics to isolate box2d code and separate particle from cell. Hopefully fixing bugs that cause box2d crash and allowing for experimenting with different physics engines

+++ core/src/com/protoevo/env/Environment.java
@@ -2,7 +2,6 @@ package com.protoevo.env;
 
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.physics.box2d.*;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.collect.Streams;
 import com.protoevo.biology.BurstRequest;
@@ -13,8 +12,8 @@ import com.protoevo.biology.cells.PlantCell;
 import com.protoevo.biology.cells.Protozoan;
 import com.protoevo.biology.evolution.Evolvable;
 import com.protoevo.core.Statistics;
-import com.protoevo.physics.Shape;
 import com.protoevo.physics.*;
+import com.protoevo.physics.box2d.Box2DPhysics;
 import com.protoevo.settings.SimulationSettings;
 import com.protoevo.utils.Geometry;
 import com.protoevo.utils.SerializableFunction;
@@ -23,7 +22,7 @@ import java.io.Serializable;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.TimeUnit;
+import java.util.stream.Stream;
 
 
 public class Environment implements Serializable
@@ -32,8 +31,7 @@ public class Environment implements Serializable
 	public static SimulationSettings settings = SimulationSettings.createDefault();
 
 	private final SimulationSettings mySettings;
-	private transient World world;
-	private float physicsStepTime;
+	private Physics physics;
 	private String loadingStatus;
 	private final Statistics stats = new Statistics();
 	private final Statistics debugStats = new Statistics();
@@ -42,7 +40,7 @@ public class Environment implements Serializable
 
 	private transient Chunks chunks;
 
-	private Map<Class<? extends Particle>, SerializableFunction<Float, Vector2>> spawnPositionFns;
+	private Map<Class<? extends Cell>, SerializableFunction<Float, Vector2>> spawnPositionFns;
 
 	private final ChemicalSolution chemicalSolution;
 	private final LightManager light;
@@ -63,7 +61,6 @@ public class Environment implements Serializable
 	private final ConcurrentHashMap<Long, Cell> cells = new ConcurrentHashMap<>();
 	private boolean hasInitialised, hasStarted;
 	private Vector2[] populationStartCentres;
-	private final JointsManager jointsManager;
 	@JsonIgnore
 	private final ConcurrentLinkedQueue<BurstRequest<? extends Cell>> burstRequests = new ConcurrentLinkedQueue<>();
 
@@ -78,8 +75,7 @@ public class Environment implements Serializable
 
 		hasStarted = false;
 		createTransientObjects();
-		buildWorld();
-		jointsManager = new JointsManager(this);
+		physics = new Box2DPhysics();
 
 		System.out.println("Creating chemicals solution... ");
 		if (Environment.settings.enableChemicalField.get()) {
@@ -102,16 +98,8 @@ public class Environment implements Serializable
 
 	public void createTransientObjects() {
 		cellsToAdd = new HashSet<>();
-//		cells = new ConcurrentHashMap<>();
 		chunks = new Chunks();
 		chunks.initialise();
-	}
-
-	public void buildWorld() {
-		world = new World(new Vector2(0, 0), true);
-		world.setContinuousPhysics(false);
-		world.setAutoClearForces(true);
-		world.setContactListener(new CollisionHandler(this));
 	}
 
 	public boolean hasStarted() {
@@ -120,39 +108,37 @@ public class Environment implements Serializable
 
 	public void rebuildWorld() {
 		settings = mySettings;
-		buildWorld();
-		createRockFixtures();
+		physics = new Box2DPhysics();
+		physics.registerStaticBodies(this);
 		for (Cell cell : getCells())
 			cell.setEnv(this);
-		jointsManager.rebuild(this);
 		updateChunkAllocations();
 	}
 
 	public void update(float delta)
 	{
 		hasStarted = true;
-		getCells().forEach(Particle::physicsUpdate);
+		getCells().forEach(cell -> cell.getParticle().physicsUpdate());
 
 		timeManager.update(delta);
 		light.update(delta);
 
-		long startTime = System.nanoTime();
-		world.step(
-				delta,
-				Environment.settings.misc.physicsVelocityIterations.get(),
-				Environment.settings.misc.physicsPositionIterations.get());
-		physicsStepTime = TimeUnit.SECONDS.convert(
-				System.nanoTime() - startTime, TimeUnit.NANOSECONDS);
+		physics.step(delta);
 
   		handleCellUpdates(delta);
 		handleBirthsAndDeaths();
 		updateChunkAllocations();
 
-		jointsManager.flushJoints();
+		physics.getJointsManager().flushJoints();
 
 		if (Environment.settings.enableChemicalField.get()) {
 			chemicalSolution.update(delta);
 		}
+	}
+
+	public void ensureAddedToEnvironment(Cell cell) {
+		if (!cells.containsKey(cell.getId()))
+			registerToAdd(cell);
 	}
 
 	private void handleCellUpdates(float delta) {
@@ -179,25 +165,7 @@ public class Environment implements Serializable
 	public void createRocks() {
 		System.out.println("Creating rocks structures...");
 		rocks.addAll(WorldGeneration.generate());
-		createRockFixtures();
-	}
-
-	public void createRockFixtures() {
-		for (Rock rock : this.getRocks()) {
-			BodyDef rockBodyDef = new BodyDef();
-			Body rockBody = world.createBody(rockBodyDef);
-			PolygonShape rockShape = new PolygonShape();
-			rockShape.set(rock.getPoints());
-			rockBody.setUserData(rock);
-
-			FixtureDef rockFixtureDef = new FixtureDef();
-			rockFixtureDef.shape = rockShape;
-			rockFixtureDef.density = 0.0f;
-			rockFixtureDef.friction = 0.7f;
-			rockFixtureDef.filter.categoryBits = ~FixtureCategories.SENSOR;
-
-			rockBody.createFixture(rockFixtureDef);
-		}
+		physics.registerStaticBodies(this);
 	}
 
 	public void initialise() {
@@ -234,8 +202,8 @@ public class Environment implements Serializable
 		spawnPositionFns = new HashMap<>(3, 1);
 		if (populationStartCentres != null) {
 			final float clusterR = Environment.settings.worldgen.populationClusterRadius.get();
-			spawnPositionFns.put(PlantCell.class, r -> randomPosition(r, populationStartCentres, 1.5f*clusterR));
-			spawnPositionFns.put(Protozoan.class, r -> randomPosition(r, populationStartCentres, clusterR));
+			spawnPositionFns.put(PlantCell.class, r -> randomPosition(r, populationStartCentres, clusterR));
+			spawnPositionFns.put(Protozoan.class, r -> randomPosition(r, populationStartCentres, 0.8f * clusterR));
 		}
 		else {
 			spawnPositionFns.put(PlantCell.class, this::randomPosition);
@@ -257,8 +225,12 @@ public class Environment implements Serializable
 		System.out.println("Creating population of " + nPlants + " plants..." );
 		loadingStatus = "Seeding Plants";
 		for (int i = 0; i < nPlants; i++) {
-			PlantCell cell = Evolvable.createNew(PlantCell.class);
-			cell.addToEnv(this);
+			PlantCell cell;
+			if (Environment.settings.plant.evolutionEnabled.get())
+				cell = Evolvable.createNew(PlantCell.class);
+			else
+				cell = new PlantCell();
+			cell.setEnv(this);
 			if (cell.isDead()) {
 				System.out.println(
 					"Failed to find position for plant pellet. " +
@@ -272,7 +244,7 @@ public class Environment implements Serializable
 		loadingStatus = "Spawning Protozoa";
 		for (int i = 0; i < nProtozoa; i++) {
 			Protozoan p = Evolvable.createNew(Protozoan.class);
-			p.addToEnv(this);
+			p.setEnv(this);
 			if (p.isDead()) {
 				System.out.println(
 					"Failed to find position for protozoan. " +
@@ -281,8 +253,8 @@ public class Environment implements Serializable
 			}
 		}
 
-		for (Particle p : cellsToAdd)
-			p.applyImpulse(Geometry.randomVector(.01f));
+		for (Cell cell : cellsToAdd)
+			cell.getParticle().applyImpulse(Geometry.randomVector(.01f));
 	}
 
 	public Vector2 randomPosition(float entityRadius, Vector2[] clusterCentres) {
@@ -302,9 +274,10 @@ public class Environment implements Serializable
 			float r = WorldGeneration.RANDOM.nextFloat() * clusterRadius;
 			Vector2 pos = Geometry.randomPointInCircle(r, WorldGeneration.RANDOM);
 			pos.add(centre);
-			Optional<? extends Shape> collision = getCollision(pos, entityRadius);
-			if (collision.isPresent() && collision.get() instanceof PlantCell) {
-				PlantCell plant = (PlantCell) collision.get();
+			Optional<? extends Shape> collision = getCollidingShape(pos, entityRadius);
+			if (collision.isPresent() && collision.get() instanceof Particle
+					&& ((Particle) collision.get()).getUserData() instanceof PlantCell) {
+				PlantCell plant = ((Particle) collision.get()).getUserData(PlantCell.class);
 				plant.kill(CauseOfDeath.ENV_CAPACITY_EXCEEDED);
 				return pos;
 			} else if (!collision.isPresent())
@@ -314,9 +287,13 @@ public class Environment implements Serializable
 		return null;
 	}
 
-	public Vector2 getRandomPosition(Particle particle) {
-		return spawnPositionFns.getOrDefault(particle.getClass(), this::randomPosition)
-				.apply(particle.getRadius());
+	public Vector2 getRandomPosition(Cell cell) {
+		return spawnPositionFns.getOrDefault(cell.getClass(), this::randomPosition)
+				.apply(cell.getRadius());
+	}
+
+	public Vector2 getRandomPosition(Class<? extends Cell> cellClass) {
+		return spawnPositionFns.getOrDefault(cellClass, this::randomPosition).apply(0f);
 	}
 
 	public Vector2 randomPosition(float entityRadius) {
@@ -344,10 +321,8 @@ public class Environment implements Serializable
 		chunks.add(cell);
 	}
 
-	public Cell getCell(long id) {
-		if (cells.containsKey(id))
-			return cells.get(id);
-		return null;
+	public Optional<Cell> getCell(long id) {
+		return Optional.ofNullable(cells.get(id));
 	}
 
 	private void flushEntitiesToAdd() {
@@ -365,17 +340,14 @@ public class Environment implements Serializable
 		getCells().forEach(chunks::allocate);
 	}
 
-	private void dispose(Particle e) {
+	private void dispose(Cell e) {
 		CauseOfDeath cod = e.getCauseOfDeath();
 		if (cod != null) {
 			int count = causeOfDeathCounts.getOrDefault(cod, 0);
 			causeOfDeathCounts.put(cod, count + 1);
 		}
-		e.dispose();
+		e.getParticle().dispose();
 	}
-
-//	private final Colour meatColourDeposit = new Colour(Color.FIREBRICK);
-//	private final Colour plantColourDeposit = new Colour(Color.FOREST);
 
 	public void depositOnDeath(Cell cell) {
 		if (settings.enableChemicalField.get()) {
@@ -383,31 +355,23 @@ public class Environment implements Serializable
 				chemicalSolution.depositCircle(
 						cell.getPos(), cell.getRadius() * 1.25f,
 						cell.getColour());
-//				if (cell instanceof Protozoan || cell instanceof MeatCell)
-//					chemicalSolution.depositCircle(
-//							cell.getPos(), cell.getRadius() * 1.5f,
-//							meatColourDeposit.cpy().mul(0.25f + 0.75f * cell.getHealth()));
-//				else if (cell instanceof PlantCell)
-//					chemicalSolution.depositCircle(
-//							cell.getPos(), cell.getRadius() * 1.5f,
-//							plantColourDeposit.cpy().mul(0.25f + 0.75f * cell.getHealth()));
 			}
 		}
 	}
 
-	public int getLocalCount(Particle cell) {
+	public int getLocalCount(Cell cell) {
 		return getLocalCount(cell.getClass(), cell.getPos());
 	}
 
-	public int getLocalCount(Class<? extends Particle> cellType, Vector2 pos) {
+	public int getLocalCount(Class<? extends Cell> cellType, Vector2 pos) {
 		return chunks.getChunkCount(cellType, pos);
 	}
 
-	public int getLocalCapacity(Particle cell) {
+	public int getLocalCapacity(Cell cell) {
 		return getLocalCapacity(cell.getClass());
 	}
 
-	public int getLocalCapacity(Class<? extends Particle> cellType) {
+	public int getLocalCapacity(Class<? extends Cell> cellType) {
 		return chunks.getChunkCapacity(cellType);
 	}
 
@@ -464,24 +428,7 @@ public class Environment implements Serializable
 	}
 
 	public Statistics getPhysicsDebugStats() {
-		debugStats.clear();
-
-		debugStats.putCount("Bodies", world.getBodyCount());
-		debugStats.putCount("Contacts", world.getContactCount());
-		debugStats.putCount("Joints", world.getJointCount());
-		debugStats.putCount("Fixtures", world.getFixtureCount());
-		debugStats.putCount("Proxies", world.getProxyCount());
-		debugStats.putTime("Physics Step Time", physicsStepTime);
-
-		int totalCells = cells.size();
-		int sleepCount = 0;
-		for (Cell cell : getCells())
-			if (cell.getBody() != null && !cell.getBody().isAwake())
-				sleepCount++;
-
-		debugStats.putPercentage("Sleeping",  100f * sleepCount / totalCells);
-
-		return debugStats;
+		return physics.getDebugStats();
 	}
 
 	public Statistics getProtozoaSummaryStats(
@@ -520,15 +467,15 @@ public class Environment implements Serializable
 		return generationCounts.getOrDefault(Protozoan.class, 0L);
 	}
 
-	public Optional<? extends Shape> getCollision(Vector2 pos, float r) {
-		Optional<Cell> collidingCell = Streams.concat(getCells().stream(), cellsToAdd.stream())
+	public Optional<? extends Shape> getCollidingShape(Vector2 pos, float r) {
+		Optional<Rock> collidingRock = rocks.stream().filter(rock -> rock.intersectsWith(pos, r)).findAny();
+		if (collidingRock.isPresent())
+			return collidingRock;
+
+		return Streams.concat(getCells().stream(), cellsToAdd.stream())
 				.filter(cell -> Geometry.doCirclesCollide(pos, r, cell.getPos(), cell.getRadius()))
+				.map(Cell::getParticle)
 				.findAny();
-
-		if (collidingCell.isPresent())
-			return collidingCell;
-
-		return rocks.stream().filter(rock -> rock.intersectsWith(pos, r)).findAny();
 	}
 
 	public float getElapsedTime() {
@@ -543,28 +490,20 @@ public class Environment implements Serializable
 		return rocks;
 	}
 
-	public World getWorld() {
-		return world;
+	public Physics getPhysics() {
+		return physics;
 	}
 
 	public Collection<Cell> getCells() {
 		return cells.values();
 	}
 
-	public Collection<? extends Particle> getParticles() {
-		return getCells();
-	}
-
-	public void ensureAddedToEnvironment(Particle particle) {
-		if (particle instanceof Cell) {
-			Cell cell = (Cell) particle;
-			if (!cells.containsKey(cell.getId()))
-				registerToAdd(cell);
-		}
+	public Stream<Particle> getParticles() {
+		return getCells().stream().map(Cell::getParticle);
 	}
 
 	public JointsManager getJointsManager() {
-		return jointsManager;
+		return physics.getJointsManager();
 	}
 
 	public <T extends Cell> void requestBurst(Cell parent,
@@ -598,7 +537,7 @@ public class Environment implements Serializable
 	}
 
 	public void incrementCrossOverCount() {
-		crossoverEvents++;
+		crossoverEvents = crossoverEvents + 1;
 	}
 
 	public float getRadius() {
@@ -606,7 +545,7 @@ public class Environment implements Serializable
 	}
 
 	public void dispose() {
-		world.dispose();
+		physics.dispose();
 	}
 
 	public LightManager getLightMap() {
